Rework Sides candidate UX and unify diary notes across Board and Sides
- Sides: hide abandoned songs from the available-song bank; give
  candidate/placed songs a shared blue treatment (badge on rows in the
  ABCD columns, solid background in the available list, no separate
  star icon); add a one-click "Mark as Candidate" toggle in the notes
  modal instead of requiring a drag-in/drag-out round trip.
- Diary notes: extract SongNotesModal/NotesButton out of the Sides
  page into components/SongNotes.tsx and reuse it on Board, replacing
  the old per-stage, create-only "+ diary entry" button with a
  persistent, labeled "Diary" button in the header that shows full
  history for the active song from any stage.
- Board: fix the song mix player, which depended on the server's
  volatile in-memory active-song state and broke whenever the API
  process restarted; switch to the song-scoped mix endpoints that
  resolve straight from disk.
- Sides diary entries always attribute to "gabriel" (single-operator
  workflow) instead of a free-text author field.
