+++ packages/client/app/sides/page.tsx
@@ -14,10 +14,10 @@ import {
   useSensors,
 } from "@dnd-kit/core";
 import {
-  assignSongToSide, createDiaryEntry, fetchDiaryEntries, fetchSides, fetchSongMixInfo,
-  fetchSongs, moveSongBetweenSides, removeSongFromSide, songMixStreamUrl,
+  assignSongToSide, fetchSides, fetchSongs, moveSongBetweenSides, removeSongFromSide,
 } from "@/lib/api";
-import { DiaryEntry, SideEntry, SideName, SideSong, SidesResponse, SongEntry } from "@/lib/types";
+import { SideEntry, SideName, SideSong, SidesResponse, SongEntry } from "@/lib/types";
+import { NotesButton, SongNotesModal } from "@/components/SongNotes";
 
 const SIDE_NAMES: SideName[] = ["A", "B", "C", "D"];
 
@@ -81,200 +81,6 @@ interface DragPayload {
   fromSide: SideName | null;
 }
 
-function SongNotesModal({ song, onClose }: { song: SongEntry; onClose: () => void }) {
-  const [mixInfo, setMixInfo] = useState<{
-    has_mix: boolean;
-    mix_file: string | null;
-    duration_seconds: number | null;
-  } | null>(null);
-  const [entries, setEntries] = useState<DiaryEntry[]>([]);
-  const [loadingEntries, setLoadingEntries] = useState(true);
-  const [author, setAuthor] = useState("");
-  const [phase, setPhase] = useState("");
-  const [title, setTitle] = useState("");
-  const [body, setBody] = useState("");
-  const [saving, setSaving] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-
-  useEffect(() => {
-    let cancelled = false;
-    setMixInfo(null);
-    setLoadingEntries(true);
-
-    fetchSongMixInfo(song.id).then((info) => {
-      if (!cancelled) setMixInfo(info);
-    });
-    fetchDiaryEntries(song.production_slug)
-      .then((fetched) => {
-        if (!cancelled) setEntries([...fetched].reverse());
-      })
-      .catch(() => {
-        if (!cancelled) setEntries([]);
-      })
-      .finally(() => {
-        if (!cancelled) setLoadingEntries(false);
-      });
-
-    return () => {
-      cancelled = true;
-    };
-  }, [song.id, song.production_slug]);
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setSaving(true);
-    setError(null);
-    try {
-      const created = await createDiaryEntry(song.production_slug, {
-        song_slug: song.production_slug,
-        author,
-        phase: phase || null,
-        title: title || null,
-        body,
-        tags: [],
-        metadata: {},
-      });
-      setEntries((prev) => [created, ...prev]);
-      setAuthor("");
-      setPhase("");
-      setTitle("");
-      setBody("");
-    } catch (err) {
-      setError(err instanceof Error ? err.message : "Save failed");
-    } finally {
-      setSaving(false);
-    }
-  };
-
-  return (
-    <div
-      className="fixed inset-0 z-50 flex items-center justify-center bg-black/70 backdrop-blur-sm"
-      onClick={onClose}
-    >
-      <div
-        className="relative bg-zinc-900 border border-zinc-700 rounded-xl w-full max-w-lg mx-4 flex flex-col max-h-[85vh] shadow-2xl"
-        onClick={(e) => e.stopPropagation()}
-      >
-        <div className="flex items-center justify-between px-5 py-3.5 border-b border-zinc-800 flex-shrink-0">
-          <span className="text-sm font-semibold text-white font-sans truncate">{song.title}</span>
-          <button
-            onClick={onClose}
-            className="text-zinc-500 hover:text-zinc-200 text-lg leading-none transition-colors"
-            aria-label="Close"
-          >
-            ×
-          </button>
-        </div>
-
-        <div className="px-5 py-3 border-b border-zinc-800/60 flex-shrink-0">
-          {mixInfo === null ? (
-            <p className="text-[11px] font-sans text-zinc-600 italic">Loading mix…</p>
-          ) : mixInfo.has_mix ? (
-            <audio controls src={songMixStreamUrl(song.id)} className="w-full h-9" />
-          ) : (
-            <p className="text-[11px] font-sans text-zinc-600 italic">No mix file yet</p>
-          )}
-        </div>
-
-        <div className="flex-1 overflow-y-auto px-5 py-3 flex flex-col gap-2">
-          {loadingEntries ? (
-            <p className="text-[11px] font-sans text-zinc-600 italic">Loading entries…</p>
-          ) : entries.length === 0 ? (
-            <p className="text-[11px] font-sans text-zinc-600 italic">No diary entries yet</p>
-          ) : (
-            entries.map((entry) => (
-              <div key={entry.id} className="border border-zinc-800 rounded px-3 py-2">
-                <div className="flex items-center justify-between gap-2 text-[10px] font-sans text-zinc-500">
-                  <span>
-                    {entry.author}
-                    {entry.phase ? ` · ${entry.phase}` : ""}
-                  </span>
-                  <span>{new Date(entry.created_at).toLocaleString()}</span>
-                </div>
-                {entry.title && (
-                  <p className="text-xs font-sans font-semibold text-zinc-200 mt-1">{entry.title}</p>
-                )}
-                <p className="text-xs font-sans text-zinc-300 mt-1 whitespace-pre-wrap">{entry.body}</p>
-              </div>
-            ))
-          )}
-        </div>
-
-        <form onSubmit={handleSubmit} className="flex flex-col gap-3 px-5 py-4 border-t border-zinc-800 flex-shrink-0">
-          <div className="flex gap-3">
-            <label className="flex flex-col gap-1 flex-1">
-              <span className="text-[10px] font-sans text-zinc-500 uppercase tracking-wider">Author</span>
-              <input
-                type="text"
-                value={author}
-                onChange={(e) => setAuthor(e.target.value)}
-                required
-                placeholder="gabriel"
-                className="bg-zinc-800 border border-zinc-700 rounded px-2.5 py-1.5 text-xs font-sans text-zinc-200 placeholder-zinc-600 focus:outline-none focus:border-zinc-500 transition-colors"
-              />
-            </label>
-            <label className="flex flex-col gap-1 flex-1">
-              <span className="text-[10px] font-sans text-zinc-500 uppercase tracking-wider">Phase</span>
-              <input
-                type="text"
-                value={phase}
-                onChange={(e) => setPhase(e.target.value)}
-                className="bg-zinc-800 border border-zinc-700 rounded px-2.5 py-1.5 text-xs font-sans text-zinc-200 placeholder-zinc-600 focus:outline-none focus:border-zinc-500 transition-colors"
-              />
-            </label>
-          </div>
-          <label className="flex flex-col gap-1">
-            <span className="text-[10px] font-sans text-zinc-500 uppercase tracking-wider">Title</span>
-            <input
-              type="text"
-              value={title}
-              onChange={(e) => setTitle(e.target.value)}
-              placeholder="optional headline"
-              className="bg-zinc-800 border border-zinc-700 rounded px-2.5 py-1.5 text-xs font-sans text-zinc-200 placeholder-zinc-600 focus:outline-none focus:border-zinc-500 transition-colors"
-            />
-          </label>
-          <label className="flex flex-col gap-1">
-            <span className="text-[10px] font-sans text-zinc-500 uppercase tracking-wider">Body</span>
-            <textarea
-              value={body}
-              onChange={(e) => setBody(e.target.value)}
-              required
-              rows={4}
-              placeholder="what happened?"
-              className="bg-zinc-800 border border-zinc-700 rounded px-2.5 py-1.5 text-xs font-sans text-zinc-200 placeholder-zinc-600 focus:outline-none focus:border-zinc-500 resize-none transition-colors"
-            />
-          </label>
-          {error && <p className="text-[10px] font-sans text-red-400">{error}</p>}
-          <button
-            type="submit"
-            disabled={saving}
-            className="w-full py-2 text-sm font-sans font-semibold rounded-lg bg-violet-700 hover:bg-violet-600 text-white disabled:opacity-40 disabled:cursor-not-allowed transition-colors"
-          >
-            {saving ? "Saving…" : "Save entry"}
-          </button>
-        </form>
-      </div>
-    </div>
-  );
-}
-
-function NotesButton({ title, onOpen }: { title: string; onOpen: () => void }) {
-  return (
-    <button
-      type="button"
-      onClick={(e) => {
-        e.stopPropagation();
-        onOpen();
-      }}
-      aria-label={`Listen and add notes for ${title}`}
-      title="Listen / diary notes"
-      className="w-[34px] h-[34px] shrink-0 flex items-center justify-center border border-[var(--ef-gray)] bg-transparent text-[#c9c9c9] hover:text-[var(--ef-orange)] hover:border-[var(--ef-orange)] transition-[color,border-color] duration-500 ease-in-out"
-    >
-      <i className="fa-solid fa-feather-pointed text-[15px]" aria-hidden="true" />
-    </button>
-  );
-}
-
 function AvailableSongRow({
   song,
   onOpenNotes,
@@ -283,6 +89,7 @@ function AvailableSongRow({
   onOpenNotes: (songId: string) => void;
 }) {
   const disabled = !song.has_mix;
+  const considered = song.lp_consideration !== "not_considered";
   const { attributes, listeners, setNodeRef, isDragging } = useDraggable({
     id: `avail:${song.id}`,
     data: { songId: song.id, title: song.title, fromSide: null } satisfies DragPayload,
@@ -296,7 +103,9 @@ function AvailableSongRow({
       className={`px-3 py-2 rounded border text-xs font-sans flex items-center justify-between gap-2 ${
         disabled
           ? "border-zinc-800 text-zinc-600 cursor-not-allowed"
-          : "border-zinc-700 text-zinc-200 cursor-grab hover:border-zinc-500"
+          : considered
+            ? "border-blue-700 bg-blue-800 text-white cursor-grab hover:border-blue-500"
+            : "border-zinc-700 text-zinc-200 cursor-grab hover:border-zinc-500"
       } ${isDragging ? "opacity-40" : ""}`}
       title={disabled ? "No mix file — cannot be sequenced" : "Drag onto a side"}
     >
@@ -341,8 +150,8 @@ function SideSongRow({
       ref={setRefs}
       {...attributes}
       {...listeners}
-      className={`px-3 py-2 rounded border bg-zinc-900 text-xs font-sans text-zinc-200 flex items-center justify-between gap-2 cursor-grab ${
-        isOver ? "border-blue-500" : "border-zinc-700 hover:border-zinc-500"
+      className={`px-3 py-2 rounded border bg-blue-900/10 text-xs font-sans text-zinc-200 flex items-center justify-between gap-2 cursor-grab ${
+        isOver ? "border-blue-400" : "border-blue-800 hover:border-blue-600"
       } ${isDragging ? "opacity-40" : ""}`}
     >
       <span className="truncate">{title}</span>
@@ -533,6 +342,7 @@ export default function SidesPage() {
   );
   const available = songs
     .filter((s) => !assignedIds.has(s.id))
+    .filter((s) => s.lifecycle_status !== "abandoned")
     .filter((s) => showUnmixed || s.has_mix)
     .filter((s) => !poolSearch || s.title.toLowerCase().includes(poolSearch.toLowerCase()));
   const unmixedHiddenCount = songs.filter((s) => !assignedIds.has(s.id) && !s.has_mix).length;
@@ -630,7 +440,7 @@ export default function SidesPage() {
         (() => {
           const notesSong = songs.find((s) => s.id === notesSongId);
           return notesSong ? (
-            <SongNotesModal song={notesSong} onClose={() => setNotesSongId(null)} />
+            <SongNotesModal song={notesSong} onClose={() => setNotesSongId(null)} onUpdated={refresh} />
           ) : null;
         })()}
     </div>
